Rename `_dominates_function.py`
